Rename comparator.h to compare.h
