+++ cpp/src/arrow/CMakeLists.txt
@@ -154,7 +154,7 @@ if(ARROW_COMPUTE)
       compute/kernels/aggregate.cc
       compute/kernels/boolean.cc
       compute/kernels/cast.cc
-      compute/kernels/comparator.cc
+      compute/kernels/compare.cc
       compute/kernels/count.cc
       compute/kernels/filter.cc
       compute/kernels/hash.cc
