+++ cpp/src/arrow/compute/kernels/compare.h
